feat: restructure reader module with new components and utilities
