+++ apps/web/src/routes/kiwi/-reader/index.ts
@@ -1,5 +1,2 @@
-export { default as ReaderContents } from "./contents";
-export * from "./flip-page-button";
-export { default as ReaderPageProgress } from "./page-progress";
-export { default as ReaderProvider } from "./provider";
-export * from "./styles";
+export * from "./components";
+export * from "./utils";
